These are my final year projects.
The mobile app used by the community includes features such as making a rescue request, being a foster care for a street dog, making donations, being a sponsor for a street dog. When making a request the informer should take a picture of the dog which in need and send it, the app locates the live location. If the location is incorrect the informer can choose the correct location from the map provided. The informer also must send a small descriptive message of the current situation of the street pooch mentioning accidents or such. If the user-sponsored a dog, they could view the progress on the app. Also, the app allows making donations for the organization to help street dogs. A dog can be adopted using this app, once a dog is adopted notifications will be provided of its days of treatments or inject of medicine.